Work around Babili/React minification bug
Before this patch the production build wasn't working - if the user
clicked on a line, there was an error logged to the console. After some
debugging it appears that the source of the problem is what happens to
the React library when it is minified by Babili. Not having found a
solution to the problem, this seems to work well enough for now and
also gives the benefit of building much faster in production. However,
there is a slight performance hit on having to parse three scripts
instead of one.

See this Babili GitHub issue for details: https://git.io/vQUI1