Remove dynamic import & replace with a Webpack alias
With these changes we no longer include development-only dependencies in
the production build of the application.

Previously all possible dependencies were being included in the
development and production bundles, including those of the opposing
environment. The solution implemented here is to use `resolve.alias` to
load a relative module by the name of `development.js` or
`production.js` depending on the value of the `NODE_ENV` environment
variable.