Use Jest to run ESLint as well as Prettier and the test suite itself
This makes it a bit nicer in development, since only one command
(`yarn test`) is required before committing and pushing code.
However, it's worth noting that in the CI environment, only the
test suite is run by Jest so the linter and formatter have to be called
manually. In practice this behaviour is not undesirable, since it means
that we can exit early if one of the three fails, starting with the
linter, then the formatter and finally running the tests.