Make Babel config more explicit
This commit removes the stage-0 preset in favor of the transforms that
are required. Furthermore, in order for it to be a bit easier to read
the Babel config, all the base options are defined in the Babel section
of the package.json, and the Webpack Babel config explicitly extends it.
The reason for defining the base options in package.json is so that all
the tooling has access to the required config (other than Webpack).