Replace fs-extra with an internal module
Loading fs-extra in the ui module was unnecessarily heavy and since it
was barely being used at all in the main process module it made sense to
remove it. However, since the new internal library relies on util.promisify,
which was introduced in Node.js v8, we also have to upgrade to the
latest Electron version, which bundles Node.js v8.