Upgrade to Jest v22
This upgrade was unfortunately not quite seamless and required a tweak
in one place and the removal of a test in another. It appears that there
is something broken with how HTML elements are created in the latest
version of jsdom (which the latest Jest uses), causing some event
listeners to break the tests. For now, these can be ignored by tests, as
it was strangely enough only happening in one place.