Remove dependency on Lodash modules where possible
This reduces the bundle size considerably and removes the need for a
complex Webpack plugin integration in order to remove unnecessary modules.
Excluding defaultsDeep and throttle, Lodash was only being used for its
utility methods. Most of these utility methods are not really necessary.
If Lodash proves to be useful enough, it may be introduced at a later
time.